[feat] python-3.11 support
