+++ funidata_utils/schemas/compat/common_serializers_312.py
@@ -1,0 +1,12 @@
+# Python > 3.12 implementations
+
+
+def serialize_as_list[typevar](v: set[typevar] | list[typevar] | None) -> list[typevar] | None:
+    if v is None:
+        return None
+
+    try:
+        return list(set(v))
+    except Exception as e:
+        # If it can't be hashed to set, just return as list
+        return list(v)
